requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.4
contourpy==1.2.1
cffi==1.16.0
cryptography==42.0.8
cycler==0.12.1
dill==0.3.8
fonttools==4.53.0
gunicorn==22.0.0
h11==0.14.0
idna==3.10
instagrapi==2.1.3
kiwisolver==1.4.5
libretranslatepy==2.1.1
lxml==5.2.2
matplotlib==3.9.0
multiprocess==0.70.16
numpy==2.0.0
numpy-stl==1.4.2
packaging==24.1
pathos==0.3.2
pillow==11.0.0
pocketsphinx==5.0.3
pox==0.3.4
ppft==1.7.6.8
pyAudio==0.2.14
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
PyExecJS==1.5.1
pyparsing==3.1.2
pypiwin32==223
pySocks==1.7.1
pyTelegramBotAPI==4.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-utils==3.8.2
pyttsx3==2.90
pywin32==306
requests==2.32.3
